requirements.txt
streamlit
pandas
numpy
scikit-learn==1.2.2
statsmodels
plotly
transformers
google-generativeai